+++ src/main/java/com/chess/engine/move/CastlingMove.java
@@ -2,8 +2,10 @@ package com.chess.engine.move;
 
 import com.chess.engine.board.Board;
 import com.chess.engine.piece.Piece;
+import com.chess.engine.player.Player;
 
 import java.awt.*;
+import java.util.List;
 
 public abstract class CastlingMove extends Move{
     // king piece and king position are extended from base class (Move)
@@ -14,5 +16,21 @@ public abstract class CastlingMove extends Move{
         super(board, king, kingDestination, king.getPosition());
         this.rook = rook;
         this.rookDestination = rookDestination;
+    }
+
+    @Override
+    public Board makeMove() {
+        Board.BoardBuilder builder = new Board.BoardBuilder();
+        Player activePlayer = board.getCurrentPlayer();
+
+        builder.setActivePlayer(activePlayer.getOpponent().getColor())
+                .placePiecesExcluding(board.getCurrentPlayer()
+                        .getOpponent().getAvailablePieces(), null)
+                .placePiecesExcluding(board.getCurrentPlayer().getAvailablePieces(), List.of(movedPiece, rook));
+
+        builder.placePiece(movedPiece.movePiece(destination));
+        builder.placePiece(rook.movePiece(rookDestination));
+
+        return builder.build();
     }
 }
