+++ src/main/java/com/chess/engine/player/WhitePlayer.java
@@ -4,6 +4,7 @@ import com.chess.engine.board.Board;
 import com.chess.engine.move.CastlingMove;
 import com.chess.engine.move.KingSideCastling;
 import com.chess.engine.move.Move;
+import com.chess.engine.move.QueenSideCastling;
 import com.chess.engine.piece.Color;
 import com.chess.engine.piece.King;
 import com.chess.engine.piece.Piece;
@@ -42,13 +43,23 @@ public class WhitePlayer extends Player {
         List<Move> castlingMoves = new ArrayList<>();
         King king = getKing();
 
-        if (king.isFirstMove() && !isInCheck() && board.getTile(7, 7).isOccupied()
-                && board.getTile(7, 7).getPiece() instanceof Rook rook
-                && rook.isFirstMove()) {
+        if (king.isFirstMove() && !isInCheck()) {
 
-            if (isTileSafeToGo(board.getTile(7, 6).getPosition()) && isTileSafeToGo(board.getTile(7, 5).getPosition())) {
+            if (board.getTile(7, 7).isOccupied()
+                    && board.getTile(7, 7).getPiece() instanceof Rook rook
+                    && rook.isFirstMove()
+                    && isTileSafeToGo(board.getTile(7, 6).getPosition())
+                    && isTileSafeToGo(board.getTile(7, 5).getPosition())) {
                 CastlingMove kingSideCastle = new KingSideCastling(board, king, board.getTile(7, 7).getPiece(), new Point(7, 6), new Point(7, 5));
                 castlingMoves.add(kingSideCastle);
+            }
+            if (board.getTile(7, 0).isOccupied()
+                    && board.getTile(7, 0).getPiece() instanceof Rook rook
+                    && rook.isFirstMove()
+                    && isTileSafeToGo(board.getTile(7, 3).getPosition())
+                    && isTileSafeToGo(board.getTile(7, 2).getPosition())) {
+                CastlingMove queenSideCastle = new QueenSideCastling(board, king, board.getTile(7, 0).getPiece(), new Point(7, 2), new Point(7, 3));
+                castlingMoves.add(queenSideCastle);
             }
         }
         return castlingMoves;
